Create InfoGeoModel.py
This is a simple feedforward neural network with 2 fully connected (fc) layers. The first layer takes the input data x with input_size features and maps it to a hidden representation of hidden_size features. The second layer maps this hidden representation to the output num_classes classes. The ReLU activation function is applied after the first fully connected layer to introduce non-linearity in the model.